Updated 3.4.13
Version Information added to WalkTheWeb Downloads - When searching for WalkTheWeb downloads, we now added version information to the feed from the 3dnet.walktheweb.com hub, as well as providing update user and date separate than the create user and date. Then we added them to the display when searching downloads from within WalkTheWeb Open-Source.

Added Int value for Version Order - Avatars, Communities, Buildings, and Things now have a version. In addition, we added a versionorder int field to the database which is a calculated sum of 1,000,000 x first digit, 1,000 x second digit, and the third digit. This makes it easier to sort the table in version order. 

Connect Web files Added Userid - On the connect communities, community, buildings, building, things, and thing, they were returning the createdate, but we added the updatedate, updateuserid, and createuserid.

Save Template Share Form Version - The Process to save the Share form for communities, buildings, and things now saves the Version and Version Description. This includes the Form fields, JavaScript functions, Handlers, and Class functions.

Read Only Version on Info Page - Since the Version is controlled by Shared Webs (set when you share an upgraded Community, Building, Thing, or Avatar),  we made the information page under Options and Settings show the Version and Latest Version Description as read only. Use the Share Page in the menu to create a new version.

Version Control in Database - For consistency, we are now using versionid, version, versionorder, and versiondesc fields to determine and track versions. Versionid is the key field value of the first version and remains the same during updated versions to keep the chain of updates preserved. Version is the 3 integers separated by period, where the first is major updates or rewrites, second is minor updates, and third is bug fixes or small modifications and adjustments. The versionorder is a numerical representation of the version used to sort the versions, as the sum of the first value multiplied by 1,000,000, second by 1,000, and third value as is. The versiondesc is a short note describing the update, Initial Version is the default value. This now applies to Avatars, Buildings, Communities, Things, UploadObjects, and UserAvatars. Some existing fields like versionuploadobjectsid were renamed to versionid for consistency.

keyExists Function Table Prefix - In the DB Class, there is a function to check if the keyExists that works with the getNewKey function. The getNewKey automatically added the tableprefix value to the table name, while the keyExists function did not. For consistency, the keyExists function now automatically appends the tableprefix value and it has been removed from the table names when calling the function. 

Version Update for User Avatars - When you load your User Avatars, the selection window will now show you if there is an updated version of your avatar! Just click to update. This should preserve your custom colors and add any missing animations. It will replace animations that match the animationevent with the updated avatar animationevent. So you should make a backup copy of any avatar changes and custom animations you have from the original, you will need to update them again after the version update. 

New User Update Avatar Process - In the Handlers, Avatar Class, and corresponding JavaScript (wtw_login.js and 3dinternet plugin), we added a process to update the user avatar to the latest version of the avatar based on the 3dnet.walktheweb.com hub repository. This includes downloading the avatar and animations. The process will overwrite any custom changes in the files, so make a backup before proceeding.

Version Check Service Now Available - We now have a versioncheck.php page available on https://3dnet.walktheweb.com/connect/ that you can send an array of items to check and receive the results of any updates currently available. Just send an array of webtype, webid, version, and versionid to receive updates. Webtype is community, building, thing, or avatar; webid is the id of the item to be updated, versionid is the id of the original version, and version is the current version loaded.