Updated 3.4.14
Stop Propagation on Update Buttons - When selecting the Update Buttons on Communities, Buildings, Things, Avatars, and User Avatars, the update button would then load the item after it updates because the button is stacked on the load button. We added the event Stop Propagation function to the process to make sure it only completes the update when clicked.