Updated 3.4.14
Stop Propagation on Update Buttons - When selecting the Update Buttons on Communities, Buildings, Things, Avatars, and User Avatars, the update button would then load the item after it updates because the button is stacked on the load button. We added the event Stop Propagation function to the process to make sure it only completes the update when clicked.

New Server Stats and Information - On the Admin Dashboard (with Admin Permission on the server) we added a list of server based stats as found in the database and file system. The connect/dashboard.php page has been rewritten to supply the data, init class changed the form for what was the 3D Websites totals, and changed the JavaScript to format the results set for the listing. 

Update Badges - Improved the Check for Updates process to include 3D Communities, 3D Buildings, 3D Things, and 3D Avatars. We added the update badges to the Admin Menu Items to show there are updates. This also improved the check for updates process by removing a duplicate call to the check function. 

Added Updates List to the main Updates page - In the Admin Menu, under Dashboard, Updates, we added a listing of all 3D Webs (Communities, Buildings, Things, and Avatars) that have updates available. This makes it a one stop place to see and perform all available updates. The list only appears when there are updates available. 

Additional Version Check Information - To simplify the results lists of updates, we passed through the webname, webdesc, and webimage with the items to have updates to make it easy to display the title, description, and thumbnail image in the list to be updated. This page comes from the 3dnet.walktheweb.com hub server.

New dirSize PHP Function available - We added a new function to get a directory size in bytes. This dirSize function is found in the DB Class file and then extended into the Handlers, connect, and plugins classes. 

Feedback page has Badges - The Feedback page now has badges to show how many NEW feedback items are listed. It also adds them to the total dashboard badge.

Pending Downloads Badges - In the Admin Menu, Dashboard page, if there are any pending WalkTheWeb Downloads, they will now trigger badge update notifications to help you find them. Download or Cancel the items to remove the badges.