Updated 3.4.14
Avatar Animation Update Process - The avatar Update process was causing unwanted duplicates for some of the animations. We improved the process to make sure duplicates do not happen. If it is an event like onwalk, onjump, onwalkbackwards, etc... it will only allow one event and the animation finds it and updates the single event. When it is an onoption event (used for unlimited animations to be loaded) it now checks for the actual animation file name to see if it is a duplicate. This will now allow the Friendly Names to be changed without creating duplicates. (formerly it was using the friendly name and file name combo to determine duplicates.)

Upgraded to BabylonJS 5.0 Game Engine! - We updated the BabylonJS Game Engine to 5.0! Check out all of the enhancements. https://www.babylonjs.com

Basic Avatars Bone Parenting - After the update to Babylon version 5, while loading an avatar, when we try to parent the bones to the avatar base it throws an error. Although we previously used this method, it may not need this step, so it has been removed from active avatars and when editing an avatar in admin mode. 

Sorted the Optional Avatar Animations - Added a sort array code to the additional animations so that you can search alphabetically to trigger the animation for your Avatar.

Load Avatar for Plugins vs Designer - When loading the avatar in the designer, the load avatar process tries to load the pluginsAvatarLoadComplete function, which is not available unless in full load mode. We added a check to see if the plugins function is available before trying to load it. Now it works when the avatar is loaded into the scene, but not when the avatar is loaded into the designer.

Save Avatar from Designer was Throwing Error - When saving your avatar from the designer, it was throwing an error on the defaultspeedratio, which was removed from the useravataranimations database table. We removed the reference to the field and now it functions without error.