Updated 3.5.0
Invalid reference to loadactionzoneid corrected - There were numerous references to loadactionzoneid in the WTW.actionZones array that were mislabeled as loadactionzone (missing the letters id). This was causing issues while updating in Admin that if the Form was opened, the new value was not being used. All reloads from the Database were using the correct property name. This was corrected in numerous places including plugin examples.

Added Spawn Action Zone ID to Table - We added a field spawnactionzoneid as a varchar(16) to the actionzones table to track an optional specific spawn zone for a teleport. 

Avoid Duplicate Water Layer - With the addition of teleport zones, communities can now be loaded after another one was loaded. So the code to add Water may be run multiple times (which was not the case previously). So new code was added to remove any existing water before the new water is created for the new scene.

Action Zone Connect Files now support Spawn Zones - The connect files for Action Zones (actionzones.php, actionzone.php, and actionzonesbywebid.php) now include the spawn zone id field (spawnzoneid) in the reply. This allows us to set and teleport to a specific Spawn Zone to enter a 3D Community. 

Added addUnloadConnectingGridToQueue Function - The addUnloadConnectingGridToQueue function is designed to queue the removal of a Connecting Grid and all items under it. This is for a future enhancement to allow multiple 3D Communities in a 3D Scene.

Create Account Blank Fields Error - The Create Account process had an error that was allowing it to try to create an account when one of the required values were blank. The process now checks and throws an appropriate error under these conditions.

Set Start Position Note Update - When you set the Default Start Position of a 3D Community, 3D Building, or 3D Thing, we added a note about the random distance within 10 units from the point (keeps avatars from stacking o the entry point) and a note about adding additional spawn zones under the Add / Edit Action Zones - edit 3D Web (Community, Building, or Thing).