Updated 3.5.1
Default Water Level in Admin - With the new Water settings and Wave Height, the default water level in admin mode was -1 and the waves were showing when ground level was 0. Now we set the default water level to -50 so that it does not show until you intentionally set it. Note that when you are not in admin mode, the water does not get added unless it should be shown. 

Spawn Action Zone ID added to Tables - The spawnactionzoneid has been added to the Communities, Buildings, and Things tables. This will be used to set the default spawn point for the web. If left blank and there are multiple spawn points, it will use a random spawn point. Note that there is a default set point (under options and settings) in addition to the added spawn action zones.

Set Default Spawn Zone - Under Admin Options and Settings (under 3D Community, 3D Building, or 3D Thing), the Set Start Position has been expanded to include setting a particular Spawn Zone as the first time entry. This is shown in the Advanced Options section. When left blank, it will randomly pick the spawn zone. The Default selection will force the Set Position from the above section to be used as the Start Position. 

New Global Variable for Default Spawn Zone ID - We added a new Global Variable called WTW.spawnZoneID that identifies the default spawn zone for the loaded 3D Community, 3D Building, or 3D Thing. When blank, it randomly selects between the default set position and any existing spawn zones (WTW.spawnZones array). The key value of default will force the 3D Scene to use the Set Start Position.

New Save Default Spawn Zone Function - The new saveDefaultSpawnZone function can be found in the class_wtwcommunities.php functions. It is used to save the spawn action zone id for 3D Communities, 3D Buildings, and 3D Things.  

New Case for savedefaultspawnzone in Handlers  - We added a new case for savedefaultspawnzone in the Communities Handler file. It also now passes the buildingid and thingid with the communityid to handle all three table updates.

Default Spawn Zone ID in Domain Info Connect - We added the default spawnactionzoneid to the domaininfo.php file (which provides info about a 3D Web not loaded yet) and the equivalent domain info provided from the session init under the main wtw class (which gets populated to the WTW.SpawnZoneID global value).

Delta Create Table Error Corrected - The deltaCreateTable php function was throwing an error for an undefined offset 2. Turns out that when the first line of the statement did not have a parenthesis with the data type it would trip the error. This was happening on the menuitems and errorlog tables.  We added code to validate if it does or does not, then construct the table update properly.  

Added Archive Date to Error Log Table - We added an archive date to the error log table so that we can separate new errors from old saved (likely resolved) errors.

New Error Log Viewer - Under Dashboard in Admin, there is a new menu option for Error Log. There are 3 views including Active Errors, Most Recent Errors, and All Errors. You can also click Delete Archived Errors to permanently delete the archived errors out of the table. Active and Most Recent will not show archived errors. Most Recent will limit the results to the latest 25 errors. Errors display the Path, File Name, Function, then the Error.

Delete Archived Error Log Function - There is now a JavaScript Function for WTW.deleteArchivedErrorLog() that will remove any archived errors in the errorlog table. Archived records are the ones with an archivedate set. This also includes a Handler and php function of the same name in the tools class. 

Integrated the htaccess file into the install - We integrated the .htaccess file into the install by using the htaccess file in the root as a template and copying it to .htaccess. The .htaccess file provides the rewrite rules to redirect virtual urls to the proper paths (Mainly for plugins). This file is specifically for Apache installs and will be ignored in IIS and NginX. Windows uses the web.config provided.

FPS Box showing on Opening 3D Scene - The default of the Show FPS (frames per second) in settings was to hide the FPS. But the surrounding box was being shown by default. If you clicked on the FPS and Off again, it would correct the issue. Now we set the box to be hidden by default. 

New 3d.walktheweb.com home page - The new 3d.walktheweb.com home page now features the latest teleporting, spawn zones, Water, Wind, and Waves features! You can now land quickly on the home page and decide which 3D Website you wish to view.