Updated 3.5.3
Group 3D Model Parts - You can now group your 3D Models as parts of a larger 3D Model. This will allow us to share complete 3D Models with all of its parts in the future. To group, click the 3D Model Group button that is not part of the group, then click the one that is (or any group part), the ungrouped part will update the groupid in the uploadobjects table to match the second 3D Model groupid (second Group Button selected). When you mouse over a group button, it will highlight all of the grouped parts that are currently on the window.

New Plugin Hook for Toggle Admin Sub Menu - We added a new plugin hook to the toggleAdminSubMenu function so that you can add additional menu items that use it.

New Toggle Menu Function - We added a new toggleAdminSubMenu function that replaces a number of separate toggle menu functions (like dashboard, media library, settings, etc...). This also makes the menu options load the first sub menu item under it where applicable and close when you click it a second time. (more consistency).

The Web Domains renamed to 3D Websites - The Web Domains root admin menu item has been renamed to 3D Websites and moved farther up the menu (right below 3D Avatars). Everything above it tends to be to create 3D Websites, everything below is optional or server settings related.

Control Panel Renamed to Login and Multiplayer Settings - The Control Panel under 3D Internet has been renamed to Login and Multiplayer Settings.

Developer Tools New Listing - The Developer Tools have been rearranged. The Quick Settings items have been removed (since they were duplicates) and the lists have been moved to the Admin Menu level 1.

Reordered the loadCoins Parameters - We reordered the loadCoins parameter so that the ghost parameter is last. This makes it easier to be an optional parameter.

Admin role Override - The checkUpdateAccess and checkAdminAccess functions have been updated to also check the Admin or Developer Role to override the granular permissions.

Added Server Franchise ID to Connect Files - We added serverfranchiseid as a blank value to many connect files. When it is a true franchise from another server, this value will be populated. This defines when a record set returns from another server or local. The serverfranchiseid is used as the first part of the mold name (blank is replaced with local).

Show and Hide Action Zone Updated - The showActionZone and hideActionZone functions have been updated to support the franchised Action Zones. Note the names of the action zones start with the text local or the server franchise id to show where they are hosted.

Parent and Child Server Franchise Values Saved - We added parentserverfranchiseid and childserverfranchiseid to the passed values and saved to the connecting grids function and handler for the connecting grids class.

Admin Hidden Fields Added for Server Franchise - We added wtw_tparentserverfranchiseid and wtw_tchildserverfranchiseid to the hidden fields in wtwadmin class. This will allow us to edit connecting grids and pass or edit values for these fields.

Server Franchise added to Object Definitions - We added parentserverfranchiseid and childserverfranchiseid to the connecting grid object definition.

Connect Connecting Grids added Server Franchise - The parentserverfranchiseid and childserverfranchiseid are now added to the connect/connectinggrids.php call. This will identify the server that the franchise is hosted.

Added Parent and Child Server Franchise ID - We added parentserverfranchiseid and childserverfranchiseid as varchar(32) to the fields on the connectinggrids table. This will assist with tracking and connecting 3D Webs from one server to another.

Select Avatar for Edit Issue - The Select Avatar for Edit in admin mode had a quote issue (from the last update). Corrected it and now it works as expected.

Added Snapshot URL to Molds - We added name, snapshotid, and snapshoturl to the /connect/moldsbywebid.php. This will allow us to show an image of the 3D Community, 3D Building, or 3D Thing that is being loaded.

Added Snapshot URL to Action Zones - We added name, snapshotid, snapshoturl, and thing analyticsid to the /connect/actionzones.php, /connect/actionzone.php, and /connect/actionzonesbywebid.php. This will allow us to show an image of the 3D Community, 3D Building, or 3D Thing that is being loaded.

Added Snapshot URL to Connecting Grids - We added snapshoturl to the /connect/connectinggrids.php. This will allow us to show an image of the 3D Community, 3D Building, or 3D Thing that is being loaded.

Permanently Delete Images or Files - You can now permanently delete images and files in my images under the Media Library in Admin Mode. When you delete an image, it will replace any use of that image with a solid dark gray stock image to avoid any orphaned images or textures in the 3D Scenes. This will delete the record in the uploads table and the actual files (original, websize, and thumbnail) if they exist.