Updated 3.5.3
Initial Plugins Activated - Some plugins were being activated automatically during updates, but were not during new installs. Added code to activate the right initial plugins.

Group ID added to Initial Table Data - The initial table data for uploadobjects now includes a default groupid. Also, the table field groupid was set to not null and is now removed to allow blank data to be written and replaced as needed.

Added Group ID to Upload Object Download Webs - We added groupid to the uploadobjects insert for WalkTheWeb Downloads. By default, if groupid does not exist, it is the uploadobjectid.

Set Old Version before Initial Data Load - The old version was being set before the table update process, but it was not set before the initial table data load process. We added a default setting before the table creation process.

Error in Initial Load of Data - We corrected the initial load of data. When loading the avatars table, the insert values were trying to insert a blank into a date field. It has been replaced with a null.

Missing Global Variable for OldVersion - There was a missing global variable $wtw->oldversion in the checked in version. It is used to test what updates are required since a particular version.

Added Get Upload Object Animation Count - We added a PHP function for getUploadObjectAnimationsCount which helps decide if the upload object has animations. This is part of the uploads class.

Added Animated to 3D Models List - We added the word Animated on the 3D Models that have animations. This makes it easy to see which have animations. An animation count is not necessary since one animation can move everything.

Added Folder Size to 3D Models List - We added the Folder Size to the 3D Models Listing. This makes it easier to compare similar 3D Models.

Added File Count to 3D Models List - We added the File Count to the 3D Models Listing. This makes it easier to compare similar 3D Models.

Added Get File Count Function - We added a getFileCount($zdirectory) function to the wtwdb and extended it to the handlers, connect, and plugins classes.

3D Model Group Listing - We added a 3D Models Group Listing that is shown on the Edit Page. This helps find grouped 3D Models and dependencies (especially as your list of 3D Models grows).

Added Clear Button to Name Filter - We added a Clear X button to the Name Filter on the Media Library Page. This makes it easier to clear and focus on a single click.

Merge Duplicate 3D Models - 3D Models now show a Duplicate button. When you select it, it filters the 3D Models to show any duplicate 3D Model names, if found, you can select the Duplicate Button on one of them to transfer all of the instances of the first 3D Model to the second 3D Model. After this is done, you can click Edit on the first 3D Model and Delete or Permanently Delete it if you choose.

Added Confirmation to Permanently Delete 3D Model - We added a confirmation process before Permanently Deleting a 3D Model.

Corrected Misspelled Permanent - The terms permanent and permanently were misspelled in a number of places in the code. It has been corrected.

Options to Delete 3D Models - You can now delete or permanently delete 3D Models as long as they are not added to any 3D Web (3D Community, 3D Building, or 3D Thing). When you Edit the 3D Model in the Media Library, You will either see a list of 3D Webs the 3D Model is shown, or you will see the options to delete or permanently delete the 3D Model. Delete will flag the record as deleted and not show in the listing any more. Permanently Delete will delete the record and checks to see if any other 3D Models are using the same folder of files (common if it was downloaded and used multiple times in a 3D Web, like multiple chairs at a table). If this is the only record using the folder, it will delete the files and folder. Note that it only deletes files with the same extensions that are allowed to be uploaded (safety precaution).

Corrected Dual Scrollbars - On the Media Library pages, there was a second scrollbar next to the first on the right side of the page. It was related to the top section buttons and div heights. we added a dynamic setting of the full page div where we were already setting the scroll section div.

List of 3D Webs using a 3D Model - We added a list of 3D Webs that are using the 3D Model when you click Edit on the 3D Model. (not visible when selecting the 3D model, just when you open media library and view 3D Models directly). You can click on one of the 3D Webs to load it into the Admin Mode Editor.

Added 3D Webs Count to 3D Models - We added the 3D Webs Count to the 3D Model listings so that you will know how many 3D Communities, 3D Buildings, and 3D Things are currently using the 3D Model.