Updated 3.5.5
Columns Selection on Downloads Page - We added columns selection on the WalkTheWeb Downloads pages. Select 1 column for large images, 2, 3, or 4 to show more items at a time.

New ID for New From Scratch - To simplify the code and make it easy to track New From Scratch items, the primary key IDs on the WalkTheWeb Downloads has been changed to communityid 0000000000000000, buildingid 1111111111111111, and thingid 2222222222222222.

Version ID Update Fix - We added a correction on the versionid in the communities, buildings, and things tables to update the versionid to the primary key id where it matched the previous New From Scratch item. It was causing any version update to trigger all of them with an unrelated update. This caused additional items to be loaded into the 3D Scene. This is now corrected.

Validating Version ID - When Downloading a new 3D Web, if the item is New from Scratch, we have to update the version id to make sure it is no longer related to the Scratch item. We added code to check before entering the record on download. If it is a New from Scratch, the versionid becomes the newwebid.

New Building from Scratch was Missing - The New Building from Scratch was missing. This was related to the Version ID of the Vintage Building being the same as the New Building From Scratch. Once we reset the versionid to a new id, the versioning of the blank worked as expected.

Upgrade 3D Webs having extra Meshes - Each 3D Web has a version ID that matches the original key ID field value. When you creates something New From Scratch, the values were still all the same causing various versions to be available for upgrade over it. In many cases this would add extra action zones and meshes to your 3D Scene. We added some code to the function for getNewKey so that it will check if it is one of the 3 New From Scratch 3D Webs and automatically assign it a new ID. Then each New From Scratch will have no other version available for upgrade.

Added Optional Upgrades Records - We added the first 4 optional upgrades. These include checking if 3D Internet is turned on (free) and Multiplayer for Admin or Hosts role will see custom domain name and SSL options.

Check Optional Upgrades Function - We added a PHP checkOptionalUpgrades function that uses the optionalupgrades table to define what to check. If you mark an item deleted in the table, it will not be checked.

Optional Upgrades in Admin Menu - If any upgrades or invoices apply, you will see Upgrades with or without the Invoices menu. This option opens the Optional Upgrades page. Currently you might see multiplayer, custom domain names, and SSL upgrades when available. We will be adding more possible upgrades as we develop this section.

Added Optional Upgrades Applied Table - We added a table called optionalupgradesapplied that stores the instances of upgrades per user, domain, ssl cert, server, or host.

Added Optional Upgrades Table - We added optionalupgrades table to track the kind of upgrades, title, description, and instructions available. It helps trigger the actual Upgrade Notices.

Store Images Not Showing - Corrected - In some cases, the store images from WooCommerce were not showing in WalkTheWeb. The issue was the images is passed as a querystring and sometimes the existing image includes a querystring adding a second question mark to the string. We used a replace to convert the ? to &. This fixed the issue. In addition, the path to the image loader has been updated to use walktheweb in the path. Both paths will work, but this is the preferred paths for plugin related API calls for WordPress.