Updated 3.5.5
Format Number function added to Install - The latest changes and additions to the download script required us to add the formatNumber function to the wtw_install.js file. This function is in the utilities file, but that is not loaded during the install. the install is a scaled down version of utilities.

Added Mime Types to web.config - For the Windows IIS hosted sites, we added the required Mime Types for babylon, obj, glb, and gltf files. This will cover the loading of 3D Models used by default in Babylon.

Add Plugin Buttons - We added 2 add plugin buttons that take you to the WalkTheWeb download plugins page. This same page can be found under Media Library -> WalkTheWeb Downloads -> 3D Plugins.

Added Image to Plugins List - We added an image to the list of plugins. The image can be added to your plugins by adding an image file of type png, jpg, or gif with the name of your plugin (example: wtw-3dinternet.png) in the root of your plugin folder.

Moved 3D Hosting Settings - We moved 3D Hosting Settings form under the settings menu to the new 3D Hosting menu item.

Created 3D Hosting Plugin - Since not everyone will use WalkTheWeb to host 3D Websites for others, we moved the 3D Hosting functionality and settings to a new 3D Hosting Plugin.

Added Plugin Hook for Open Domain Form - We added a plugin Hook for the openDomainForm function. This allows us to add 3D Hosting settings to the process.

Reformatted Create Wizard Columns - In the 3D Create Wizard (Pedestal Button), we reformatted the page to allow the user to select the number of columns shown as well as reorder the results to have the title, image, then additional information.

Added Animation to Create Website Button - On the new Create 3D Website Pedestal we added animation to the button.

CSG Update Issue - When editing a Mold, if you select to subtract the mold from another mold, the CSG process was not immediately updating the scene. In addition, when you tried to reload the CSG settings, it was deleting the subtraction, causing it to save without the subtraction. The setNewMold function is now run after the process to update the screen and the line clearing the selected action was removed from the setMoldFormFields function. It is now working as expected.