Updated 3.6.0
Added New 3D User Menu - We added in a new User Menu that is in-game, 3D and interactive. This will make it more available for all devices (our end goal).

Added Animated Loading 3D Object - We added an animated (rotating) 3D Object to signify loading. It can display text for loading 3D Avatar, Loading 3D Model, Loading 3D Scene, Loading 3D Community, Loading 3D Building, Loading 3D Thing, and Loading. Currently we have it in use for Loading Avatar. This makes it easier to know when the avatar loading and downloading is taking place.

3D Internet Plugin Removed settings - The following 3 settings (wtw3dinternet_enableGlobal, wtw3dinternet_enableLocal, wtw3dinternet_enableAnonymous) are now replaced by the WTW global settings: WTW_globalLogins, WTW_localLogins, and WTW_anonymousLogins. This allowed us to remove the previous settings from the 3D Internet Plugin and load settings function.

Added Loading Global Login Variables - We added a function to load the Login Global Variables. This function takes place during the core loading, which allows the variables to be used at login time. Variables are: WTW.globalLogins, (settings value WTW_globalLogins), WTW.localLogins (settings value WTW_localLogins), and WTW.anonymousLogins (settings value WTW_anonymousLogins).

Added Global Variables for Login Types - The Login Types: globalLogins, localLogins, and anonymousLogins global JavaScript variables have been moved from the 3D Internet Plugin Class to the main WTW class. This will make it easier to test in any code regardless if the 3D Plugin is active or not.

Email mislabeled as User Email - In the wtw users class, sql mentioning the users table was referencing useremail instead of email as a field name. This has been corrected.

Save Profile Email Correction - In the saveProfile and saveMyProfile functions in the wtw Users class, the process to check for duplicates email was catching your own email as a duplicate and not updating it. We added code to check the current user email and save the profile, correcting the issue.

Added 3dforms Script - We added a 3dforms JavaScript page under the /core/scripts/molds folder. This script will assist with 3D Objects being treated as forms. It is similar to the 3D blog page and the hud fields script.

Prevent Object not found in Login - Based on how the login is called, the wtw and wtwuser objects may or may not be initiated. We added code to check if they are set before attempting to add the user id. This replaces the try and catch exception process for setting them.

Correction to Rewrite Rule - In the htaccess file, there is a correction on the rewrite rule from [P] to [L} which stands for Last and stop processing any lines after the match and this rewrite is made. The [P] turned out to be Apache specific. Now the rule is more global.

Added Mime Types to the htaccess file - We added the 6 common mime types to the htaccess file. Note that this file is used as a template to create the .htaccess file during install. You may need to manually update your .htaccess file.

Select Avatar List Correction - After a Local Login, the select Avatar list was showing numerous global avatars to choose that were not necessarily tied to the current user. The list was finding incomplete users off the global server. A correction to the SQL on the global server call corrected the issue.

Added Keyboard Filtering to key-down - Under the input keydown processing when a WTW.selectedMoldName is active, we added filtering the allowed keys based on type of input in the name of the selected mold. Currently we support Name, Email, Search, password, and default.

Added Checkbox and Button key-down - Under the input keydown processing when a WTW.selectedMoldName is active, we added support for any key to trigger checking or clearing a checkbox or triggering a button push.

Added Tab Keyboard key-down - Under the input keydown processing when a WTW.selectedMoldName is active, we added a section to catch a tab that will move from a selected mold name to the next available based on the WTW.tabOrder and if the mold name actually exists in the current scene.

Added HUD Login Script Page - We added a new HUD Login page under /core/scripts/hud folder called wtw_hud_login.js. This page is used for the new 3d Login 3D Objects and functionality.

Global Variables for 3D Forms - We added a number of global variables to be used with 3D Forms and the WTW.selectedMoldName global variable. They include: WTW.offsetX, WTW.offsetY, WTW.offsetZ, WTW.maxLength, WTW.textTimer, WTW.tabOrder, and WTW.webStyle. The Offsets are used to make the 3D Text appear on a surface of the intended 3D Object. Max Length is used to limit the number of shown characters, where they scroll in place to show the end of the text. Text Timer is for the blinking text and refreshing the text changes. Tab Order will allow tab to move to the next Selected Mold Name. Web Style sets the text properties.

Has Value Functions Pass Through - The function for hasValue() was a pass-through function actually found in the wtwdb class. The variable is passed by reference and not by value. But for a pass through that type of function call is depreciated. So I removed the pass-through by copying the function to the handler, connect, and plugin classes.

New Install Table Data - When doing a new install, the table data for the help menu was not being added. This was caused by an else condition on the if statement. The else has been removed allowing the section of data updates to be performed properly.

New Reset Password - We added a new Reset Password in-game, 3D, interactive object. This will eventually replace the HTML based overlay screen form.

New Create Login - We added a new Create Login in-game, 3D, interactive object. This will eventually replace the HTML based overlay screen form. This screen covers the basics. The additional, optional Profile fields will be added in a different 3D Object at a later date.

Login Object has Multiple Configurations - The WalkTheWeb Login 3D Object also has alternate configurations to double as the Local login in-game, 3D, interactive Object.

New WalkTheWeb Global Login - We added a new WalkTheWeb Global Login in-game, 3D, interactive object. This will eventually replace the HTML based overlay screen form.

Added new Login Menu - We added in a new Login Menu that is in-game, 3D and interactive. This will make it more available for all devices (our end goal).

Added Simple Color Covering - The Add Covering Color function has been expanded to also allow single color selection. You can set color to a HTML color code or all four colors as: color.diffuseColor, color.emissiveColor, color.specularColor, and color.ambientColor.