Updated 3.6.1
New Load Building and Community - The new install default option uses the loadInitBuildingCommunity PHP function that can be found in class_wtwtables.php. This function provides the records needed for the default 3D Building, 3D Community, and 3D Things. It is only processed if you choose default install. Note that the files associated with these 3D Objects are now included in the WalkTheWeb GitHub download.