Updated 3.6.1
Added Upload Max Size Variables - We added three hidden fields to make the upload_max_filesize, post_max_size, and memory_limit easily available in code. These values have been converted from M to byte by adding 6 zeros and multiplying it by 1.024..

Updates to File Permissions - There are a number of file permissions updates. First, wtw_chmod is now defaulting to 775 and umask is defaulting to 0. We are now also disabling the umask during the chmod setting. The wtw_umask variable is only used to reset the umask after the file permissions are set. These new settings should help with permissions groups causing issues on some platforms. These settings can still be manually changed in the /config/wtw_config.php or set under Admin mode, Settings -> Server Settings menu option.

New Load Building and Community - The new install default option uses the loadInitBuildingCommunity PHP function that can be found in class_wtwtables.php. This function provides the records needed for the default 3D Building, 3D Community, and 3D Things. It is only processed if you choose default install. Note that the files associated with these 3D Objects are now included in the WalkTheWeb GitHub download.

Columns Selection on Install - Under the Custom Install options, we added a column selection to make it easier to view and compare multiple selections. This is now more consistent with the WTW Downloads screens under admin mode (under Media Library).

Reversed Install Order - Under the new Custom option on install, we changed the order to start by selecting the 3D Building before the 3D Community. This now matches the Create 3D Community in-scene process order.

New Install Method Choice - During the install, there is now a choice to use the default or custom install. Default uses a 3D Building and 3D Community that is already downloaded and distributed with the software. It also uses a new function to load the tables for the default 3D Building, 3D Community, and related 3D Things. Custom install will allow you to select which 3D Building and 3D Community to download as your first 3D Scene.

Add first Building Step Disabled - During the download process that gives us the progress bar, the step to add the first building has been disabled (process set to false). This has been handled by another method. The issue was that if the community being downloaded has a building in it, the building would be added a second time in the center of the community. The new method avoids this issue altogether by selecting the building first, then selecting the community to add the building into the community.

Display Name Correction - The display name was not being updated to the bottom menu bar. There were two places that needed to be updated. The first was in the class_wtw-initsession.php file where the email was being set and the displayname was missing. The second was in the saveMyProfile JavaScript function where the displayname was not being updated after a change was made to the database. Both have been corrected.

Data issue with moving 3D Things - There was a number of reported issues with moving 3D Models, 3D Building Blocks, and 3D Things when editing. Turns out that the old code for version upgrades (3D Communities, 3D Buildings, and 3D Things) had caused data corruption. The code to move objects is correct. We just need to check the data in the table. The cause of the version issue was corrected in previous versions. We will double check the available WTW Downloads for corrupt data.

Duplicate Keys on 3D Forms Corrected - The duplicate keys on 3D Forms was caused by the shopping plugin. We added a condition that checks for productsearch as the field so that the shopping code only works on the right fields (4 functions needed this correction). This corrected the issue.

Avatar Drops in Water While Loading - Sometimes when the avatar is loaded faster than the Land under the avatar, the avatar will drop with gravity to the bottom of the water. We previously added a startstand which is an invisible box under the avatar to prevent this action. The box was set to disappear in 10 seconds. But there are times that the land may take longer than 10 seconds. Therefore, I changed the code to see if the init cycle is completed after 10 seconds. If not, it checks again when you move the avatar to see if the startstand is still in the scene and removes it.

Corrected 3D Form Focus for empty Forms - When the 3D Form is empty (not set to remember login) the form focus was still being set to the button. Corrected this by setting it to the email.

Corrected wtw-tinyimg CSS Class - Corrected the wtw-tinyimg CSS class to be img. specific. This corrected the icons for the downloads pages and column sets.